+++ HISTORY.md
@@ -1,1 +1,12 @@
 # history
+
+## 0.1.1
+
+### Bug fix
+
+* value 的 size 大于 options 的 size 报错
+* defaultValue 与 options 异步生成时无法渲染的错误
+
+## 0.1.0
+
+init